# services/booking-service/src/main/java/com/example/booking/api/dto/CreateBookingRequest.java
package com.example.booking.api.dto;

import jakarta.validation.Valid;
import jakarta.validation.constraints.NotBlank;
import jakarta.validation.constraints.NotEmpty;
import jakarta.validation.constraints.NotNull;

import java.util.List;

public record CreateBookingRequest(
    @NotBlank String offerId,
    @NotNull @Valid ContactDto contact,
    @NotEmpty @Valid List<PassengerDto> passengers
) {}