# services/offers-service/src/main/java/com/example/offers/api/dto/FlightOfferSummary.java
package com.example.offers.api.dto;

import java.time.Instant;

public record FlightOfferSummary(
    String id,
    String fromIATA,
    String toIATA,
    Instant departAt,
    Instant arriveAt,
    Money price,
    String carrier,
    Instant validUntil
) {}